+++ esp_audio_codec/src/audio_encoder_reg.c
@@ -52,6 +52,10 @@ esp_audio_err_t esp_audio_enc_register_default(void)
 #ifdef CONFIG_AUDIO_ENCODER_LC3_SUPPORT
     ret |= esp_lc3_enc_register();
 #endif /* CONFIG_AUDIO_ENCODER_LC3_SUPPORT */
+
+#ifdef CONFIG_AUDIO_ENCODER_G722_SUPPORT
+    ret |= esp_g722_enc_register();
+#endif /* CONFIG_AUDIO_ENCODER_G722_SUPPORT */
     return ret;
 }
 
@@ -97,4 +101,8 @@ void esp_audio_enc_unregister_default(void)
 #ifdef CONFIG_AUDIO_ENCODER_LC3_SUPPORT
     esp_audio_enc_unregister(ESP_AUDIO_TYPE_LC3);
 #endif /* CONFIG_AUDIO_ENCODER_LC3_SUPPORT */
+
+#ifdef CONFIG_AUDIO_ENCODER_G722_SUPPORT
+    esp_audio_enc_unregister(ESP_AUDIO_TYPE_G722);
+#endif /* CONFIG_AUDIO_ENCODER_G722_SUPPORT */
 }