+++ esp_audio_codec/src/simple_decoder_reg.c
@@ -22,6 +22,10 @@ esp_audio_err_t esp_audio_simple_dec_register_default(void)
 #ifdef CONFIG_AUDIO_SIMPLE_DEC_TS_SUPPORT
     ret |= esp_ts_dec_register();
 #endif /* CONFIG_AUDIO_SIMPLE_DEC_TS_SUPPORT */
+
+#ifdef CONFIG_AUDIO_SIMPLE_DEC_OGG_SUPPORT
+    ret |= esp_ogg_dec_register();
+#endif /* CONFIG_AUDIO_SIMPLE_DEC_OGG_SUPPORT */
     return ret;
 }
 
@@ -38,4 +42,8 @@ void esp_audio_simple_dec_unregister_default(void)
 #ifdef CONFIG_AUDIO_SIMPLE_DEC_TS_SUPPORT
     esp_ts_dec_unregister();
 #endif /* CONFIG_AUDIO_SIMPLE_DEC_TS_SUPPORT */
+
+#ifdef CONFIG_AUDIO_SIMPLE_DEC_OGG_SUPPORT
+    esp_ogg_dec_unregister();
+#endif /* CONFIG_AUDIO_SIMPLE_DEC_OGG_SUPPORT */
 }