+++ esp_audio_codec/test_apps/audio_codec_test/install.sh
@@ -1,4 +1,3 @@
-rm -rf sdkconfig sdkconfig.old managed_components dependencies.lock build
-idf.py set-target esp32s3
-export IDF_EXTRA_ACTIONS_PATH=managed_components/espressif__esp_board_manager
-idf.py gen-bmgr-config -b esp32_s3_korvo2_v3
+pip install esp-bmgr-assist
+pip install --upgrade esp-bmgr-assist
+idf.py bmgr -l